fix: resolve all -Werror warnings across library, tests, and benchmarks
- Sign-conversion: int→size_t for all array/vector indexing loops
- Shadow: rename local P1/P2 → pt1/pt2 where shadowing fe52_constants
- Unused variables: [[maybe_unused]] or (void) for intentionally unused
- Unused functions: remove dead h32() after sscanf→manual hex migration
- Deprecated: sscanf→manual hex parsing (MSVC CRT portability)
- Int→uint64_t: explicit static_cast in from_uint64() call sites
- Add codecov.yml configuration
- Rewrite CI coverage job to use LLVM source-based coverage

Build: cmake --build build-werror -j 8 (zero warnings with -Werror -Wall -Wextra -Wpedantic -Wconversion -Wshadow)
Tests: 12/12 passed (ctest --test-dir build-werror)

+++ cpu/bench/bench_adaptive_glv.cpp
@@ -43,7 +43,7 @@ static double run_bench(unsigned window_bits, bool glv, unsigned iters) {
         (void)p;
     }
     auto end = std::chrono::high_resolution_clock::now();
-    double ns = std::chrono::duration_cast<std::chrono::nanoseconds>(end - start).count();
+    double ns = static_cast<double>(std::chrono::duration_cast<std::chrono::nanoseconds>(end - start).count());
     return ns / static_cast<double>(iters);
 }
